+++ src/Init.cu
@@ -2,21 +2,21 @@
 
 extern int NRAD, NSEC, size_grid;
 
-extern double *Label_d;
-extern double *Rmed,  *Rmed_d;
+extern float *Label_d;
+extern float *Rmed,  *Rmed_d;
 
 extern dim3 dimGrid2, dimBlock2;
 
-__host__ void Initialization (double *Dens, double *Vrad, double *Vtheta, double *Energy, double *Label, PlanetarySystem *sys)
+__host__ void Initialization (float *Dens, float *Vrad, float *Vtheta, float *Energy, float *Label, PlanetarySystem *sys)
 {
   InitEuler (Vrad, Vtheta, Dens, Energy);
   InitLabel (Label, sys);
   WriteDim();
 }
 
-__host__ void InitLabel (double *Label, PlanetarySystem *sys)
+__host__ void InitLabel (float *Label, PlanetarySystem *sys)
 {
-  double xp,yp, rhill, rp;
+  float xp,yp, rhill, rp;
   xp = sys->x[0];
   yp = sys->y[0];
   rp = sqrt(xp*xp+yp*yp);
